*Created many developers and QAs for projects

+++ app/controllers/projects_controller.rb
@@ -2,14 +2,18 @@
 
 # project controller
 class ProjectsController < ApplicationController
-  before_action :authorization, only: %i[index new create edit update destroy]
-  before_action :project_find, only: %i[edit update show destroy]
+  #before_action :authorization, only: %i[index new create edit update]
+  before_action :set_project, only: %i[edit update show destroy]
 
   def index
-    @projects = if current_user.manager? || current_user.qa?
+
+    @projects = if current_user.manager?
+                  current_user.projects.page(params[:page]).per(6)
+                elsif current_user.qa?
                   Project.page(params[:page]).per(6)
                 else
-                  Project.page(params[:page]).per(6).where(dev_id: current_user.id)
+                  puts :dev_id
+                  Project.page(params[:page]).per(6).where(":dev_id = ANY(dev_id)", dev_id: current_user.id)
                 end
   end
 
@@ -18,7 +22,7 @@ class ProjectsController < ApplicationController
   end
 
   def edit
-
+    authorize @project
   end
 
   def create
@@ -26,11 +30,12 @@ class ProjectsController < ApplicationController
     if @project.save
       redirect_to @project, flash: { notice: 'Post Created Successfully' }
     else
-      render new_project_path
+      render new_project_path , flash: { notice: 'Something went wrong!' }
     end
   end
 
   def destroy
+    authorize @project
     @project.users.clear
     if @project.destroy
       redirect_to projects_path, flash: { notice: 'Post Deleted Successfully' }
@@ -40,10 +45,21 @@ class ProjectsController < ApplicationController
   end
 
   def show
-    @bugs = Project.includes(:bugs).where(id: params[:id]).map(&:bugs).flatten
+    if current_user.manager? || current_user.developer?
+      authorize @project
+    end
+    @bugs = @project.bugs
+    if params[:new]
+      @bugs = @project.bugs.New
+    elsif params[:started]
+      @bugs = @project.bugs.where(bug.New)
+    elsif params[:completed]
+      @bugs = Bug.all
+    end
   end
 
   def update
+    authorize @project
     if @project.update(project_params)
       redirect_to @project
     else
@@ -52,16 +68,15 @@ class ProjectsController < ApplicationController
   end
 
   private
+    def authorization
+      authorize @project
+    end
 
-  def project_params
-    params.require(:project).permit(:name, :description, :dev_id)
-  end
+    def set_project
+      @project = Project.find(params[:id])
+    end
 
-  def authorization
-    authorize Project
-  end
-
-  def project_find
-    @project = Project.find(params[:id])
-  end
+    def project_params
+      params.require(:project).permit(:name, :description, dev_id: [])
+    end
 end
